requirements.txt
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
pyyaml==6.0.1
loguru==0.7.2
openpyxl==3.1.4
rich==13.7.1
streamlit==1.35.0
pandas==2.2.2
